requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
requests==2.32.2
urllib3==2.2.1
